implementing modal....need to finish up later

+++ assets/javascript/login.js
@@ -28,32 +28,43 @@ $(document).ready(function(){
             console.log($fname);
             console.log($lname);
             console.log($username);
-     
-            // var usernameVal = $username.val();
-            currentUser = $username.val().trim();
-           
-            // add an entry for the new user and set attributes to fname and lname
-            database.ref("users/"+currentUser).set({
-                    fname: $fname.val().trim(),
-                    lname: $lname.val().trim()
-            });
+
+            if ($fname.val() !== "" && $lname.val() !== "" && $username.val() !== ""){
+
+                // var usernameVal = $username.val();
+                currentUser = $username.val().trim();
+            
+                // add an entry for the new user and set attributes to fname and lname
+                database.ref("users/"+currentUser).set({
+                        fname: $fname.val().trim(),
+                        lname: $lname.val().trim()
+                });
 
 
-            // Set the username in Session Storage
-            if(currentUser !== ""){
-                // sessionStorage.clear();
-                // sessionStorage.setItem("username", currentUser);
-                localStorage.clear();
-                localStorage.setItem("username", currentUser);
-               }
+                // Set the username in Session Storage
+                if(currentUser !== ""){
+                    // sessionStorage.clear();
+                    // sessionStorage.setItem("username", currentUser);
+                    localStorage.clear();
+                    localStorage.setItem("username", currentUser);
+                }
 
-            // Clear the input boxes
-            $fname.val("");
-            $lname.val("");
-            $username.val("");
+                // Clear the input boxes
+                $fname.val("");
+                $lname.val("");
+                $username.val("");
 
-            // Navigate to the main content page             
-            window.location.href = 'content.html';
+                // Navigate to the main content page             
+                window.location.href = 'content.html';
+            }
+            else{
+                $("#myModal").show();
+
+            }
+
+            
+
+
 
             
         });
